more minimum
It identifies all feasible point with the minimum number of selected  point (i.e all the minimum point if Q is the identity matrix) during A generation. This work.

It tries to generate Q (either diagonal or upper triangular) solving a linear system in order to set the objective of a given number of the previous extracted point to the same in value, in order to have more than one global minimum. However, it does not always work. Maybe some check on the "independence" of the extracted point should be done.